+++ go/services/manager/ServiceActivate.go
@@ -6,7 +6,7 @@ import (
 	"github.com/saichler/l8types/go/ifs"
 )
 
-func (this *ServiceManager) Activate(typeName string, serviceName string, serviceArea uint16,
+func (this *ServiceManager) Activate(typeName string, serviceName string, serviceArea byte,
 	r ifs.IResources, l ifs.IServiceCacheListener, args ...interface{}) (ifs.IServiceHandler, error) {
 
 	if typeName == "" {
